feat: refactor answer service and introduce common prompt tools
